feat: adicionando a biblioteca no projeto e tambem implementando as paginas
