+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.46.1",
+    "react-router-dom": "^6.16.0",
     "yup": "^1.2.0"
   },
   "devDependencies": {
