+++ src/pages/forgot-password.tsx
@@ -1,0 +1,5 @@
+export const ForgotPassword = () => {
+    return (
+        <p>teste</p>
+    )
+}
